Starting Debugging Voronoi Namespace
Added UnitTests project to solution
Made various bug fixes

+++ src/MapGeneration/Util.cpp
@@ -2,6 +2,6 @@
 
 using namespace Nimbus::Voronoi;
 
-bool Util::closeEnough(float d1, float d2, float diff){
+bool Util::closeEnough(double d1, double d2, double diff){
 	return abs(d1 - d2) <= diff;
 }//end closeEnough
